+++ Implementacija/MigrantControlSystem/MigrantControlSystem/Models/Lokacija.cs
@@ -9,7 +9,7 @@ namespace MigrantControlSystem.Models
     public class Lokacija
     {
         public int id { get; set; }
-        public int x { get; set; }
-        public int y { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
     }
 }
